Moved implementation from ArrayStorage to AbstractArrayStorage
