+++ src/com/urise/webapp/storage/ArrayStorage.java
@@ -2,77 +2,13 @@ package com.urise.webapp.storage;
 
 import com.urise.webapp.model.Resume;
 
-import java.util.Arrays;
-
 /**
  * Array based storage for Resumes
  */
 public class ArrayStorage extends AbstractArrayStorage {
-    private static final int STORAGE_LIMIT = 10_000;
 
-    private Resume[] storage = new Resume[STORAGE_LIMIT];
-    // Количество резюме
-    private int size = 0;
-
-    public void clear() {
-        Arrays.fill(storage, 0, size, null);
-        size = 0;
-    }
-
-    public void update(Resume resume) {
-        if (resume == null) return;
-
-        int indexToUpdate = getIndex(resume.getUuid());
-        if (indexToUpdate == -1) {
-            System.out.println("No resume with uuid: " + resume.getUuid());
-            return;
-        }
-
-        storage[indexToUpdate] = resume;
-    }
-
-    public void save(Resume resume) {
-        if (resume == null) return;
-
-        if (size >= STORAGE_LIMIT) {
-            System.out.println("Storage overflow.");
-            return;
-        }
-        if (getIndex(resume.getUuid()) != -1) {
-            System.out.println("Resume with uuid " + resume.getUuid() + " already exists.");
-            return;
-        }
-
-        storage[size] = resume;
-        size++;
-    }
-
-    public Resume get(String uuid) {
-        if (uuid == null) return null;
-
-        int index = getIndex(uuid);
-        if (index == -1) {
-            System.out.println("No resume with uuid: " + uuid);
-            return null;
-        }
-        return storage[index];
-    }
-
-    public void delete(String uuid) {
-        if (uuid == null) return;
-
-        int indexToDelete = getIndex(uuid);
-        if (indexToDelete == -1) {
-            System.out.println("No resume with uuid: " + uuid);
-            return;
-        } else {
-            storage[indexToDelete] = storage[size - 1];
-            storage[size - 1] = null;
-            size--;
-        }
-    }
-
-    private int getIndex(String uuid) {
+    @Override
+    protected int getIndex(String uuid) {
         for (int i = 0; i < size; i++) {
             if (uuid.equals(storage[i].getUuid())) {
                 return i;
@@ -81,16 +17,13 @@ public class ArrayStorage extends AbstractArrayStorage {
         return -1;
     }
 
-    /**
-     * @return array, contains only Resumes in storage (without null)
-     */
-    public Resume[] getAll() {
-        Resume[] allResume = new Resume[size];
-        System.arraycopy(storage, 0, allResume, 0, size);
-        return allResume;
+    @Override
+    protected void insertElement(Resume resume, int index) {
+        storage[size] = resume;
     }
 
-    public int size() {
-        return size;
+    @Override
+    protected void fillDeletedElement(int index) {
+        storage[index] = storage[size - 1];
     }
 }
