+++ src/com/urise/webapp/storage/SortedArrayStorage.java
@@ -10,12 +10,12 @@ public class SortedArrayStorage extends AbstractArrayStorage {
     }
 
     @Override
-    public void update(Resume r) {
+    public void update(Resume resume) {
         throw new NotImplementedException();
     }
 
     @Override
-    public void save(Resume r) {
+    public void save(Resume resume) {
         throw new NotImplementedException();
     }
 
@@ -31,6 +31,21 @@ public class SortedArrayStorage extends AbstractArrayStorage {
 
     @Override
     public Resume[] getAll() {
+        throw new NotImplementedException();
+    }
+
+    @Override
+    protected int getIndex(String uuid) {
+        throw new NotImplementedException();
+    }
+
+    @Override
+    protected void insertElement(Resume resume, int index) {
+        throw new NotImplementedException();
+    }
+
+    @Override
+    protected void fillDeletedElement(int index) {
         throw new NotImplementedException();
     }
