+++ src/com/urise/webapp/storage/Storage.java
@@ -6,9 +6,9 @@ public interface Storage {
 
     void clear();
 
-    void update(Resume r);
+    void update(Resume resume);
 
-    void save(Resume r);
+    void save(Resume resume);
 
     Resume get(String uuid);
